requirements.txt
amqp==5.0.6
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
Deprecated==1.2.13
ecdsa==0.17.0
fastapi==0.70.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
joblib==1.1.0
kombu==5.2.2
motor==2.5.1
numpy==1.21.4
pandas==1.3.4
passlib==1.7.4
prompt-toolkit==3.0.24
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pymongo==3.12.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==4.0.2
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.16.0
threadpoolctl==3.0.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
